Refactor: 팔로워/팔로잉 페이지의 API 분리
